python/requirements.txt
mysql-connector-python<=8.0.19
Flask<=1.0.2
python-dotenv<=0.14.0